+++ backend/canisters/community/impl/src/updates/accept_p2p_swap.rs
@@ -2,42 +2,43 @@ use crate::activity_notifications::handle_activity_notification;
 use crate::timer_job_types::NotifyEscrowCanisterOfDepositJob;
 use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_tracing_macros::trace;
-use chat_events::ReserveP2PTradeResult;
-use community_canister::accept_p2p_trade_offer::{Response::*, *};
+use community_canister::accept_p2p_swap::{Response::*, *};
 use ic_cdk_macros::update;
 use icrc_ledger_types::icrc1::transfer::TransferError;
-use types::{ChannelId, Chat, MessageId, MessageIndex, UserId};
+use types::{AcceptSwapSuccess, ChannelId, Chat, MessageId, MessageIndex, UserId};
 
 #[update]
 #[trace]
-async fn accept_p2p_trade_offer(args: Args) -> Response {
+async fn accept_p2p_swap(args: Args) -> Response {
     run_regular_jobs();
 
-    let ReserveP2PTradeOfferResult { user_id, c2c_args } = match mutate_state(|state| reserve_p2p_trade_offer(&args, state)) {
+    let ReserveP2PSwapResult { user_id, c2c_args } = match mutate_state(|state| reserve_p2p_swap(&args, state)) {
         Ok(result) => result,
-        Err(response) => return response,
+        Err(response) => return *response,
     };
 
-    let result = match user_canister_c2c_client::c2c_accept_p2p_trade_offer(user_id.into(), &c2c_args).await {
-        Ok(user_canister::c2c_accept_p2p_trade_offer::Response::Success(transaction_index)) => {
+    let result = match user_canister_c2c_client::c2c_accept_p2p_swap(user_id.into(), &c2c_args).await {
+        Ok(user_canister::c2c_accept_p2p_swap::Response::Success(transaction_id)) => {
             NotifyEscrowCanisterOfDepositJob::run(
                 user_id,
                 c2c_args.offer_id,
                 args.channel_id,
                 args.thread_root_message_index,
                 args.message_id,
-                transaction_index,
+                transaction_id,
             );
-            Success
+            Success(AcceptSwapSuccess {
+                token1_txn_in: transaction_id,
+            })
         }
-        Ok(user_canister::c2c_accept_p2p_trade_offer::Response::TransferError(TransferError::InsufficientFunds { .. })) => {
+        Ok(user_canister::c2c_accept_p2p_swap::Response::TransferError(TransferError::InsufficientFunds { .. })) => {
             InsufficientFunds
         }
         Ok(response) => InternalError(format!("{response:?}")),
         Err(error) => InternalError(format!("{error:?}")),
     };
 
-    if !matches!(result, Success) {
+    if !matches!(result, Success(_)) {
         mutate_state(|state| {
             rollback(
                 args.channel_id,
@@ -52,68 +53,64 @@ async fn accept_p2p_trade_offer(args: Args) -> Response {
     result
 }
 
-struct ReserveP2PTradeOfferResult {
+struct ReserveP2PSwapResult {
     user_id: UserId,
-    c2c_args: user_canister::c2c_accept_p2p_trade_offer::Args,
+    c2c_args: user_canister::c2c_accept_p2p_swap::Args,
 }
 
-fn reserve_p2p_trade_offer(args: &Args, state: &mut RuntimeState) -> Result<ReserveP2PTradeOfferResult, Response> {
+fn reserve_p2p_swap(args: &Args, state: &mut RuntimeState) -> Result<ReserveP2PSwapResult, Box<Response>> {
     if state.data.is_frozen() {
-        return Err(ChatFrozen);
+        return Err(Box::new(ChatFrozen));
     }
 
     let caller = state.env.caller();
     if let Some(member) = state.data.members.get(caller) {
         if member.suspended.value {
-            return Err(UserSuspended);
+            return Err(Box::new(UserSuspended));
         }
         let user_id = member.user_id;
 
         if let Some(channel) = state.data.channels.get_mut(&args.channel_id) {
             let channel_member = match channel.chat.members.get(&user_id) {
                 Some(m) => m,
-                _ => return Err(UserNotInChannel),
+                _ => return Err(Box::new(UserNotInChannel)),
             };
             let now = state.env.now();
 
-            match channel.chat.events.reserve_p2p_trade(
+            match channel.chat.events.reserve_p2p_swap(
                 user_id,
                 args.thread_root_message_index,
                 args.message_id,
                 channel_member.min_visible_event_index(),
                 now,
             ) {
-                ReserveP2PTradeResult::Success(boxed) => {
-                    let result = *boxed;
+                types::ReserveP2PSwapResult::Success(result) => {
                     handle_activity_notification(state);
 
-                    Ok(ReserveP2PTradeOfferResult {
+                    Ok(ReserveP2PSwapResult {
                         user_id,
-                        c2c_args: user_canister::c2c_accept_p2p_trade_offer::Args {
-                            offer_id: result.offer_id,
+                        c2c_args: user_canister::c2c_accept_p2p_swap::Args {
+                            offer_id: result.content.offer_id,
                             chat: Chat::Channel(caller.into(), args.channel_id),
                             created: result.created,
                             created_by: result.created_by,
-                            input_token: result.input_token,
-                            input_amount: result.input_amount,
-                            input_transaction_index: result.input_transaction_index,
-                            output_token: result.output_token,
-                            output_amount: result.output_amount,
-                            expires_at: result.expires_at,
+                            token0: result.content.token0,
+                            token0_amount: result.content.token0_amount,
+                            token0_txn_in: result.content.token0_txn_in,
+                            token1: result.content.token1,
+                            token1_amount: result.content.token1_amount,
+                            expires_at: result.content.expires_at,
                         },
                     })
                 }
-                ReserveP2PTradeResult::Cancelled => Err(OfferCancelled),
-                ReserveP2PTradeResult::Expired => Err(OfferExpired),
-                ReserveP2PTradeResult::AlreadyReserved(_) => Err(AlreadyAccepted),
-                ReserveP2PTradeResult::AlreadyCompleted(_) => Err(AlreadyCompleted),
-                ReserveP2PTradeResult::OfferNotFound => Err(OfferNotFound),
+                types::ReserveP2PSwapResult::Failure(status) => Err(Box::new(StatusError(status.into()))),
+                types::ReserveP2PSwapResult::OfferNotFound => Err(Box::new(OfferNotFound)),
             }
         } else {
-            Err(UserNotInChannel)
+            Err(Box::new(UserNotInChannel))
         }
     } else {
-        Err(UserNotInCommunity)
+        Err(Box::new(UserNotInCommunity))
     }
 }
 
@@ -128,6 +125,6 @@ fn rollback(
         channel
             .chat
             .events
-            .unreserve_p2p_trade(user_id, thread_root_message_index, message_id, state.env.now());
+            .unreserve_p2p_swap(user_id, thread_root_message_index, message_id, state.env.now());
     }
 }