+++ backend/canisters/user/api/src/updates/c2c_accept_p2p_swap.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use icrc_ledger_types::icrc1::transfer::TransferError;
 use serde::{Deserialize, Serialize};
-use types::{Chat, TimestampMillis, TokenInfo, UserId};
+use types::{Chat, TimestampMillis, TokenInfo, TransactionId, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -9,17 +9,17 @@ pub struct Args {
     pub chat: Chat,
     pub created: TimestampMillis,
     pub created_by: UserId,
-    pub input_token: TokenInfo,
-    pub input_amount: u128,
-    pub input_transaction_index: u64,
-    pub output_token: TokenInfo,
-    pub output_amount: u128,
+    pub token0: TokenInfo,
+    pub token0_amount: u128,
+    pub token0_txn_in: TransactionId,
+    pub token1: TokenInfo,
+    pub token1_amount: u128,
     pub expires_at: TimestampMillis,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
-    Success(u64), // The transaction index
+    Success(TransactionId),
     TransferError(TransferError),
     InternalError(String),
 }