+++ backend/canisters/user/impl/src/lib.rs
@@ -4,7 +4,7 @@ use crate::model::direct_chats::DirectChats;
 use crate::model::group_chat::GroupChat;
 use crate::model::group_chats::GroupChats;
 use crate::model::hot_group_exclusions::HotGroupExclusions;
-use crate::model::p2p_trades::P2PTrades;
+use crate::model::p2p_swaps::P2PSwaps;
 use crate::model::token_swaps::TokenSwaps;
 use crate::timer_job_types::{RemoveExpiredEventsJob, TimerJob};
 use candid::Principal;
@@ -23,13 +23,15 @@ use types::{
     BuildVersion, CanisterId, Chat, ChatId, ChatMetrics, CommunityId, Cryptocurrency, Cycles, Document, Notification,
     TimestampMillis, Timestamped, UserId,
 };
-use user_canister::NamedAccount;
+use user_canister::{NamedAccount, UserCanisterEvent};
+use utils::canister_event_sync_queue::CanisterEventSyncQueue;
 use utils::env::Environment;
 use utils::regular_jobs::RegularJobs;
 
 mod crypto;
 mod governance_clients;
 mod guards;
+mod jobs;
 mod lifecycle;
 mod memory;
 mod model;
@@ -179,7 +181,10 @@ struct Data {
     pub saved_crypto_accounts: Vec<NamedAccount>,
     pub next_event_expiry: Option<TimestampMillis>,
     pub token_swaps: TokenSwaps,
-    pub p2p_trades: P2PTrades,
+    #[serde(default)]
+    pub p2p_swaps: P2PSwaps,
+    #[serde(default)]
+    pub user_canister_events_queue: CanisterEventSyncQueue<UserCanisterEvent>,
     pub rng_seed: [u8; 32],
 }
 
@@ -229,7 +234,8 @@ impl Data {
             saved_crypto_accounts: Vec::new(),
             next_event_expiry: None,
             token_swaps: TokenSwaps::default(),
-            p2p_trades: P2PTrades::default(),
+            p2p_swaps: P2PSwaps::default(),
+            user_canister_events_queue: CanisterEventSyncQueue::default(),
             rng_seed: [0; 32],
         }
     }