+++ backend/canisters/user/impl/src/updates/accept_p2p_swap.rs
@@ -1,20 +1,22 @@
 use crate::guards::caller_is_owner;
-use crate::model::p2p_trades::{P2PTradeOffer, P2PTradeOfferStatus};
+use crate::model::p2p_swaps::{P2PSwapOffer, P2PSwapOfferStatus};
 use crate::timer_job_types::NotifyEscrowCanisterOfDepositJob;
 use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_api_macros::update_msgpack;
 use canister_tracing_macros::trace;
-use chat_events::{ReserveP2PTradeOfferSuccess, ReserveP2PTradeResult};
 use escrow_canister::deposit_subaccount;
 use icrc_ledger_types::icrc1::account::Account;
 use icrc_ledger_types::icrc1::transfer::{TransferArg, TransferError};
-use types::{CanisterId, Chat, EventIndex, TimestampMillis, UserId};
-use user_canister::accept_p2p_trade_offer::{Response::*, *};
+use types::{
+    AcceptSwapSuccess, CanisterId, Chat, EventIndex, ReserveP2PSwapResult, ReserveP2PSwapSuccess, TimestampMillis,
+    TransactionId, UserId,
+};
+use user_canister::accept_p2p_swap::{Response::*, *};
 use utils::time::NANOS_PER_MILLISECOND;
 
 #[update_msgpack(guard = "caller_is_owner")]
 #[trace]
-async fn accept_p2p_trade_offer(args: Args) -> Response {
+async fn accept_p2p_swap(args: Args) -> Response {
     run_regular_jobs();
 
     let PrepareResult {
@@ -24,21 +26,22 @@ async fn accept_p2p_trade_offer(args: Args) -> Response {
         now,
     } = match mutate_state(|state| prepare(&args, state)) {
         Ok(ok) => ok,
-        Err(response) => return response,
+        Err(response) => return *response,
     };
 
+    let content = reserve_success.content;
     let transfer_result = match icrc_ledger_canister_c2c_client::icrc1_transfer(
-        reserve_success.output_token.ledger,
+        content.token1.ledger,
         &TransferArg {
             from_subaccount: None,
             to: Account {
                 owner: escrow_canister_id,
-                subaccount: Some(deposit_subaccount(my_user_id, reserve_success.offer_id)),
+                subaccount: Some(deposit_subaccount(my_user_id, content.offer_id)),
             },
-            fee: Some(reserve_success.output_token.fee.into()),
+            fee: Some(content.token1.fee.into()),
             created_at_time: Some(now * NANOS_PER_MILLISECOND),
             memo: None,
-            amount: (reserve_success.output_amount + reserve_success.output_token.fee).into(),
+            amount: (content.token1_amount + content.token1.fee).into(),
         },
     )
     .await
@@ -59,31 +62,32 @@ async fn accept_p2p_trade_offer(args: Args) -> Response {
 
     match transfer_result {
         Ok(index) => {
+            let token1_txn_in = TransactionId { index, hash: None };
             mutate_state(|state| {
-                state.data.p2p_trades.add(P2PTradeOffer {
-                    id: reserve_success.offer_id,
+                state.data.p2p_swaps.add(P2PSwapOffer {
+                    id: content.offer_id,
                     chat: Chat::Direct(args.user_id.into()),
                     created_by: reserve_success.created_by,
                     created: reserve_success.created,
-                    status: P2PTradeOfferStatus::Accepted,
+                    status: P2PSwapOfferStatus::Accepted,
                     last_updated: state.env.now(),
-                    input_token: reserve_success.input_token,
-                    input_amount: reserve_success.input_amount,
-                    input_transaction_index: Some(reserve_success.input_transaction_index),
-                    output_token: reserve_success.output_token,
-                    output_amount: reserve_success.output_amount,
-                    output_transaction_index: Some(index),
-                    expires_at: reserve_success.expires_at,
+                    token0: content.token0,
+                    token0_amount: content.token0_amount,
+                    token0_txn_in: Some(content.token0_txn_in),
+                    token1: content.token1,
+                    token1_amount: content.token1_amount,
+                    token1_txn_in: Some(token1_txn_in),
+                    expires_at: content.expires_at,
                 });
             });
-            NotifyEscrowCanisterOfDepositJob::run(reserve_success.offer_id);
-            Success(index)
+            NotifyEscrowCanisterOfDepositJob::run(content.offer_id);
+            Success(AcceptSwapSuccess { token1_txn_in })
         }
         Err(response) => {
             mutate_state(|state| {
                 if let Some(chat) = state.data.direct_chats.get_mut(&args.user_id.into()) {
                     let now = state.env.now();
-                    chat.events.unreserve_p2p_trade(my_user_id, None, args.message_id, now);
+                    chat.events.unreserve_p2p_swap(my_user_id, None, args.message_id, now);
                 }
             });
             response
@@ -94,31 +98,28 @@ async fn accept_p2p_trade_offer(args: Args) -> Response {
 struct PrepareResult {
     my_user_id: UserId,
     escrow_canister_id: CanisterId,
-    reserve_success: ReserveP2PTradeOfferSuccess,
+    reserve_success: ReserveP2PSwapSuccess,
     now: TimestampMillis,
 }
 
-fn prepare(args: &Args, state: &mut RuntimeState) -> Result<PrepareResult, Response> {
+fn prepare(args: &Args, state: &mut RuntimeState) -> Result<PrepareResult, Box<Response>> {
     if let Some(chat) = state.data.direct_chats.get_mut(&args.user_id.into()) {
         let my_user_id = state.env.canister_id().into();
         let now = state.env.now();
         match chat
             .events
-            .reserve_p2p_trade(my_user_id, None, args.message_id, EventIndex::default(), now)
+            .reserve_p2p_swap(my_user_id, None, args.message_id, EventIndex::default(), now)
         {
-            ReserveP2PTradeResult::Success(result) => Ok(PrepareResult {
+            ReserveP2PSwapResult::Success(reserve_success) => Ok(PrepareResult {
                 my_user_id,
                 escrow_canister_id: state.data.escrow_canister_id,
-                reserve_success: *result,
+                reserve_success,
                 now,
             }),
-            ReserveP2PTradeResult::Cancelled => Err(OfferCancelled),
-            ReserveP2PTradeResult::Expired => Err(OfferExpired),
-            ReserveP2PTradeResult::AlreadyReserved(_) => Err(AlreadyAccepted),
-            ReserveP2PTradeResult::AlreadyCompleted(_) => Err(AlreadyCompleted),
-            ReserveP2PTradeResult::OfferNotFound => Err(OfferNotFound),
+            ReserveP2PSwapResult::Failure(status) => Err(Box::new(StatusError(status.into()))),
+            ReserveP2PSwapResult::OfferNotFound => Err(Box::new(OfferNotFound)),
         }
     } else {
-        Err(ChatNotFound)
+        Err(Box::new(ChatNotFound))
     }
 }